Install the wercker binary to /usr/local/bin
It's really a shame I couldn't get wercker to (properly) work when
launched from within a docker container.

The issue I ran into was that a 'wercker build' would inevitably err out
with something similar to:

--> Running step: wercker-init
/bin/bash: line 48: /pipeline/wercker-init-101e4aa8-dbf1-41f5-a6c2-1506a85b9789/init.sh: No such file or directory
--> Step failed: wercker-init
--> Pipeline failed

stracing and a few other things did not help very much :(

This is likely related to how the my (host) docker socket gets mounted
inside the container which causes wercker to barf?

I'll revisit this again soon :)