Add the symlink needed for my docker devbox
This was the workaround I needed in order for the wercker cli to be
installed inside my devbox docker container.

The story there is that I've set '/home/dev' as $HOME inside the devbox
docker container, and since that container shares a (privileged)
filesystem with the host, it gets angry (and weird!) when /home/dev
isn't where it should be.